+++ src/Models.cs
@@ -36,11 +36,39 @@ namespace LudusaviPlaynite
         public int ProcessedBytes;
     }
 
+    public struct ApiFile
+    {
+        [JsonProperty("failed")]
+        public bool Failed;
+        [JsonProperty("bytes")]
+        public int Bytes;
+        [JsonProperty("originalPath")]
+        public string OriginalPath;
+    }
+
+    public struct ApiRegistry
+    {
+        [JsonProperty("failed")]
+        public bool Failed;
+    }
+
+    public struct ApiGame
+    {
+        [JsonProperty("decision")]
+        public string Decision;
+        [JsonProperty("files")]
+        public Dictionary<string, ApiFile> Files;
+        [JsonProperty("registry")]
+        public Dictionary<string, ApiRegistry> Registry;
+    }
+
     public struct ApiResponse
     {
         [JsonProperty("errors")]
         public ApiErrors Errors;
         [JsonProperty("overall")]
         public ApiOverall Overall;
+        [JsonProperty("games")]
+        public Dictionary<string, ApiGame> Games;
     }
 }