+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 [Ludusavi v0.7.0 or newer](https://github.com/mtkennerly/ludusavi/releases) is now required.
 
 * Added:
+  * After doing a backup or restore of all games, you can click the notification
+    for a full list of all games processed.
   * "Browse" button for Ludusavi executable file.
   * "Browse" and "open" buttons for backup directory.
 * Fixed:
@@ -17,6 +19,7 @@
   * File size units are now adjusted based on the size, rather than always using MiB.
   * The default backup directory now writes out the user folder in full rather
     than using the `~` placeholder (although that is still supported).
+  * You can now dismiss notifications by clicking on them.
 
 ## v0.1.0 (2020-07-29)